Debugged most CRUD operation errors
[+] - Debugged most CRUD errors, one still exists. Used Postman and dove deeper into 4xx status codes. Also improved the methods a bit.
[*] - saveCheckout() method still gives a 404 error.
[*] - Backend endpoints should have exceptions for IDs that doesn't exist.

+++ frontend/src/app/components/book-detail/book-detail.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { BookService } from '../../services/book.service';
 import { Book } from '../../models/book';
 import { Observable } from 'rxjs';
-import { ActivatedRoute } from '@angular/router';
+import {ActivatedRoute, Router} from '@angular/router';
 import { map, switchMap } from 'rxjs/operators';
 import {Checkout} from "../../models/checkout";
 import {CheckoutService} from "../../services/checkout.service";
@@ -24,6 +24,7 @@ export class BookDetailComponent implements OnInit {
     private route: ActivatedRoute,
     private bookService: BookService,
     private checkoutService: CheckoutService,
+    private router: Router
   ) {
   }
 
@@ -34,44 +35,46 @@ export class BookDetailComponent implements OnInit {
       .pipe(switchMap(id => this.bookService.getBook(id)))
   }
 
-  // TODO redirect user to Library/Checkouts after action
-
   // https://www.w3schools.com/jsref/met_win_confirm.asp
   deleteThisBook(book: Book): void{
     if (confirm("Do you want to delete the '" + book.title + "' book?")){
+      if (book.status === 'BORROWED'){
+        alert("Can't delete a borrowed book!")
+        return;
+      }
       console.log(book.id);
       console.log(this.bookService.getBook('ab1a8f87-272d-4ba0-93df-dbb8952909df'));
       this.bookService.deleteBook(book.id).subscribe(() => {
         alert("Book deleted successfully!");
-        window.location.reload();
+        // window.location.reload();
+        // https://stackoverflow.com/questions/38131293/angular-2-router-navigate
+        this.router.navigate(["/books"]); // ROuting user back to "/books"
       }) //Backend needs to throw an exception if id doesn't exist (currently it deletes the book and then sends a 500 error)
     }
   }
-  /*
-  zsh: frontend % curl -X DELETE 'http://localhost:8080/api/book/deleteBook?bookId=ab1a8f87-272d-4ba0-93df-dbb8952909df'
-  {"timestamp":"2023-03-23T18:50:46.313+00:00","status":500,"error":"Internal Server Error","message":"","path":"/api/book/deleteBook"}%
-  */
 
   // Might need to add hover css, so that the button indicates that the book is unavailable for checkout.
   // TODO check if user has filled the first and last name
+  // TODO this method is way too long
   checkoutThisBook(book: Book): void { //
     // Implement the name check for !empty values
     if (book.status === 'AVAILABLE') {
-      if (!this.bookService.isValidInput(this.borrowerFirstName)) {
-        alert('Invalid first name. Special characters are not allowed.');
-        return;
-      }
-      if (!this.bookService.isValidInput(this.borrowerLastName)) {
-        alert('Invalid last name. Special characters are not allowed.');
-        return;
-      }
+      // if (!this.bookService.isValidInput(this.borrowerFirstName)) {
+      //   alert('Invalid first name. Special characters are not allowed.');
+      //   return;
+      // }
+      // if (!this.bookService.isValidInput(this.borrowerLastName)) {
+      //   alert('Invalid last name. Special characters are not allowed.');
+      //   return;
+      // }
       // The confirm() method returns true if the user clicked "OK", otherwise false.
       if (confirm("Do you want to check out '" + book.title + "'?")) {
         book.status = 'BORROWED';
         book.checkOutCount += 1;
         let checkoutDate = new Date();
-        checkoutDate.setDate(checkoutDate.getDate() + 60); // kinda long...
-        book.dueDate = checkoutDate.toString().substring(0, 10); // 'yyyy-mm-dd'
+        checkoutDate.setDate(checkoutDate.getDate() + 60);
+        book.dueDate = checkoutDate.toISOString().substring(0, 10); // 'yyyy-mm-dd'
+        // https://gist.github.com/makuska/84457e0b6f614301b14575aaeaa0d917#debugging-journey
 
         // this.bookService.deleteBook(book.id).subscribe(() => console.log("Temp delete")); // Working
         const checkout: Checkout = {
